Update pixi-js-1.md
Updated to use version 5.3.3, this includes updating the js cdn links and links to documentation. I also move the script tags in the head, and changed the application creation to use a JSON object, instead of just passing in the width and height. The old way still works, but passing in JSON allows for additional options to be passed in at a later date.